Add interface for screenshots
implements `shoot()`
